backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
bitarray==3.7.2
boto3==1.29.7
botocore==1.32.7
celery==5.3.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==1.0.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.11.0
cryptography==41.0.7
cytoolz==1.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
django-ratelimit==4.1.0
django-redis==6.0.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
factory-boy==3.3.0
Faker==19.12.0
frozenlist==1.8.0
gunicorn==21.2.0
hexbytes==0.3.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
lru-dict==1.2.0
multidict==6.7.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psycopg==3.2.12
psycopg-binary==3.2.12
py-ecc==8.0.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
pyunormalize==17.0.0
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
regex==2025.10.23
requests==2.31.0
rlp==4.1.0
rpds-py==0.28.0
s3transfer==0.8.0
sendgrid==6.10.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
toolz==1.1.0
twilio==8.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.14
web3==6.11.3
websockets==15.0.1
wheel==0.45.1
whitenoise==6.6.0
yarl==1.22.0